@@ -16,7 +16,7 @@
 package com.lzy.demo.base;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 
 import java.util.ArrayList;
