@@ -18,13 +18,14 @@ package com.lzy.demo.okrx2;
 import android.Manifest;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
 import android.text.TextUtils;
 import android.text.format.Formatter;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
+
+import androidx.core.app.ActivityCompat;
 
 import com.lzy.demo.R;
 import com.lzy.demo.base.BaseRxDetailActivity;
@@ -40,7 +41,7 @@ import com.lzy.okrx2.adapter.ObservableResponse;
 import java.io.File;
 import java.text.NumberFormat;
 
-import butterknife.Bind;
+import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 import io.reactivex.Observable;
@@ -66,13 +67,13 @@ public class RxFileDownloadActivity extends BaseRxDetailActivity {
 
     private static final int REQUEST_PERMISSION_STORAGE = 0x01;
 
-    @Bind(R.id.et_url) EditText etUrl;
-    @Bind(R.id.fileDownload1) Button btnFileDownload1;
-    @Bind(R.id.fileDownload2) Button btnFileDownload2;
-    @Bind(R.id.downloadSize) TextView tvDownloadSize;
-    @Bind(R.id.tvProgress) TextView tvProgress;
-    @Bind(R.id.netSpeed) TextView tvNetSpeed;
-    @Bind(R.id.pbProgress) NumberProgressBar pbProgress;
+    @BindView(R.id.et_url) EditText etUrl;
+    @BindView(R.id.fileDownload1) Button btnFileDownload1;
+    @BindView(R.id.fileDownload2) Button btnFileDownload2;
+    @BindView(R.id.downloadSize) TextView tvDownloadSize;
+    @BindView(R.id.tvProgress) TextView tvProgress;
+    @BindView(R.id.netSpeed) TextView tvNetSpeed;
+    @BindView(R.id.pbProgress) NumberProgressBar pbProgress;
     private NumberFormat numberFormat;
 
     @Override
@@ -95,7 +96,7 @@ public class RxFileDownloadActivity extends BaseRxDetailActivity {
     }
 
     @Override
-    public void onRequestPermissionsResult(int requestCode, @android.support.annotation.NonNull String[] permissions, @android.support.annotation.NonNull int[] grantResults) {
+    public void onRequestPermissionsResult(int requestCode, @androidx.annotation.NonNull String[] permissions, @androidx.annotation.NonNull int[] grantResults) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         if (requestCode == REQUEST_PERMISSION_STORAGE) {
             if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
