@@ -16,7 +16,6 @@
 package com.lzy.demo.okdownload;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.text.format.Formatter;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -25,6 +24,8 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.lzy.demo.R;
@@ -42,7 +43,7 @@ import java.text.NumberFormat;
 import java.util.List;
 import java.util.Map;
 
-import butterknife.Bind;
+import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 
@@ -119,14 +120,14 @@ public class DownloadAdapter extends RecyclerView.Adapter<DownloadAdapter.ViewHo
 
     public class ViewHolder extends RecyclerView.ViewHolder {
 
-        @Bind(R.id.icon) ImageView icon;
-        @Bind(R.id.name) TextView name;
-        @Bind(R.id.priority) TextView priority;
-        @Bind(R.id.downloadSize) TextView downloadSize;
-        @Bind(R.id.tvProgress) TextView tvProgress;
-        @Bind(R.id.netSpeed) TextView netSpeed;
-        @Bind(R.id.pbProgress) NumberProgressBar pbProgress;
-        @Bind(R.id.start) Button download;
+        @BindView(R.id.icon) ImageView icon;
+        @BindView(R.id.name) TextView name;
+        @BindView(R.id.priority) TextView priority;
+        @BindView(R.id.downloadSize) TextView downloadSize;
+        @BindView(R.id.tvProgress) TextView tvProgress;
+        @BindView(R.id.netSpeed) TextView netSpeed;
+        @BindView(R.id.pbProgress) NumberProgressBar pbProgress;
+        @BindView(R.id.start) Button download;
         private DownloadTask task;
         private String tag;
 
