@@ -36,7 +36,7 @@ import com.lzy.demo.utils.Urls;
 import com.lzy.okgo.OkGo;
 import com.lzy.okgo.model.Response;
 
-import butterknife.Bind;
+import butterknife.BindView;
 import butterknife.ButterKnife;
 
 /**
@@ -50,7 +50,7 @@ import butterknife.ButterKnife;
  */
 public class MethodActivity extends BaseDetailActivity implements AdapterView.OnItemClickListener {
 
-    @Bind(R.id.gridView) GridView gridView;
+    @BindView(R.id.gridView) GridView gridView;
 
     private String[] methods = {"GET", "HEAD", "OPTIONS", "POST", "PUT", "DELETE", "PATCH", "TRACE"};
 
