@@ -16,7 +16,6 @@
 package com.lzy.demo.okupload;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.text.format.Formatter;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -25,6 +24,8 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.lzy.demo.R;
@@ -48,7 +49,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Random;
 
-import butterknife.Bind;
+import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 
@@ -162,14 +163,14 @@ public class UploadAdapter extends RecyclerView.Adapter<UploadAdapter.ViewHolder
 
     public class ViewHolder extends RecyclerView.ViewHolder {
 
-        @Bind(R.id.icon) ImageView icon;
-        @Bind(R.id.name) TextView name;
-        @Bind(R.id.priority) TextView priority;
-        @Bind(R.id.downloadSize) TextView downloadSize;
-        @Bind(R.id.tvProgress) TextView tvProgress;
-        @Bind(R.id.netSpeed) TextView netSpeed;
-        @Bind(R.id.pbProgress) NumberProgressBar pbProgress;
-        @Bind(R.id.upload) Button upload;
+        @BindView(R.id.icon) ImageView icon;
+        @BindView(R.id.name) TextView name;
+        @BindView(R.id.priority) TextView priority;
+        @BindView(R.id.downloadSize) TextView downloadSize;
+        @BindView(R.id.tvProgress) TextView tvProgress;
+        @BindView(R.id.netSpeed) TextView netSpeed;
+        @BindView(R.id.pbProgress) NumberProgressBar pbProgress;
+        @BindView(R.id.upload) Button upload;
         private UploadTask<?> task;
         private String tag;
 
