@@ -16,18 +16,19 @@
 package com.lzy.demo.okupload;
 
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.Button;
+
+import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.lzy.demo.R;
 import com.lzy.demo.base.BaseActivity;
 import com.lzy.okserver.OkUpload;
 import com.lzy.okserver.task.XExecutor;
 
-import butterknife.Bind;
+import butterknife.BindView;
 import butterknife.OnClick;
 
 /**
@@ -41,11 +42,11 @@ import butterknife.OnClick;
  */
 public class UploadAllActivity extends BaseActivity implements XExecutor.OnAllTaskEndListener {
 
-    @Bind(R.id.toolbar) Toolbar toolbar;
-    @Bind(R.id.recyclerView) RecyclerView recyclerView;
-    @Bind(R.id.select) Button select;
-    @Bind(R.id.upload) Button upload;
-    @Bind(R.id.deleteAll) Button deleteAll;
+    @BindView(R.id.toolbar) Toolbar toolbar;
+    @BindView(R.id.recyclerView) RecyclerView recyclerView;
+    @BindView(R.id.select) Button select;
+    @BindView(R.id.upload) Button upload;
+    @BindView(R.id.deleteAll) Button deleteAll;
 
     private OkUpload okUpload;
     private UploadAdapter adapter;
